+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 0.0.1
+## [1.0.0] - 2025-12-17
 
-* TODO: Describe initial release.
+* Initial release of the audio device detection plugin.
+* Support for detecting Bluetooth, wired headset, and speaker status.
+* Implemented for both Android and iOS.